Moves akka serializer to serializer package
